# api/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create or update KeyVal asynchroniously
	// (PUT /async)
	PutKeyvalAsync(ctx echo.Context) error
	// Get KeyVal by key from cache (inconsistency possible)
	// (GET /fast)
	GetKeyvalFast(ctx echo.Context, params GetKeyvalFastParams) error
	// Delete existing KeyVal
	// (DELETE /keyval)
	DeleteKeyval(ctx echo.Context, params DeleteKeyvalParams) error
	// Get KeyVal by key
	// (GET /keyval)
	GetKeyval(ctx echo.Context, params GetKeyvalParams) error
	// Create new KeyVal
	// (POST /keyval)
	CreateKeyval(ctx echo.Context) error
	// Update existing KeyVal
	// (PUT /keyval)
	UpdateKeyval(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PutKeyvalAsync converts echo context to params.
func (w *ServerInterfaceWrapper) PutKeyvalAsync(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutKeyvalAsync(ctx)
	return err
}

// GetKeyvalFast converts echo context to params.
func (w *ServerInterfaceWrapper) GetKeyvalFast(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetKeyvalFastParams
	// ------------- Required query parameter "key" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "key", ctx.QueryParams(), &params.Key, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter key: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetKeyvalFast(ctx, params)
	return err
}

// DeleteKeyval converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteKeyval(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params DeleteKeyvalParams
	// ------------- Required query parameter "key" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "key", ctx.QueryParams(), &params.Key, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter key: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteKeyval(ctx, params)
	return err
}

// GetKeyval converts echo context to params.
func (w *ServerInterfaceWrapper) GetKeyval(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetKeyvalParams
	// ------------- Required query parameter "key" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "key", ctx.QueryParams(), &params.Key, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter key: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetKeyval(ctx, params)
	return err
}

// CreateKeyval converts echo context to params.
func (w *ServerInterfaceWrapper) CreateKeyval(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateKeyval(ctx)
	return err
}

// UpdateKeyval converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateKeyval(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateKeyval(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.PUT(baseURL+"/async", wrapper.PutKeyvalAsync)
	router.GET(baseURL+"/fast", wrapper.GetKeyvalFast)
	router.DELETE(baseURL+"/keyval", wrapper.DeleteKeyval)
	router.GET(baseURL+"/keyval", wrapper.GetKeyval)
	router.POST(baseURL+"/keyval", wrapper.CreateKeyval)
	router.PUT(baseURL+"/keyval", wrapper.UpdateKeyval)

}

type BadRequestJSONResponse ErrorResponse

type NotFoundJSONResponse ErrorResponse

type PutKeyvalAsyncRequestObject struct {
	Body *PutKeyvalAsyncJSONRequestBody
}

type PutKeyvalAsyncResponseObject interface {
	VisitPutKeyvalAsyncResponse(w http.ResponseWriter) error
}

type PutKeyvalAsync200Response struct {
}

func (response PutKeyvalAsync200Response) VisitPutKeyvalAsyncResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type PutKeyvalAsync400JSONResponse struct{ BadRequestJSONResponse }

func (response PutKeyvalAsync400JSONResponse) VisitPutKeyvalAsyncResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetKeyvalFastRequestObject struct {
	Params GetKeyvalFastParams
}

type GetKeyvalFastResponseObject interface {
	VisitGetKeyvalFastResponse(w http.ResponseWriter) error
}

type GetKeyvalFast200JSONResponse KeyVal

func (response GetKeyvalFast200JSONResponse) VisitGetKeyvalFastResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetKeyvalFast400JSONResponse struct{ BadRequestJSONResponse }

func (response GetKeyvalFast400JSONResponse) VisitGetKeyvalFastResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetKeyvalFast404JSONResponse struct{ NotFoundJSONResponse }

func (response GetKeyvalFast404JSONResponse) VisitGetKeyvalFastResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type DeleteKeyvalRequestObject struct {
	Params DeleteKeyvalParams
}

type DeleteKeyvalResponseObject interface {
	VisitDeleteKeyvalResponse(w http.ResponseWriter) error
}

type DeleteKeyval200Response struct {
}

func (response DeleteKeyval200Response) VisitDeleteKeyvalResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type DeleteKeyval400JSONResponse struct{ BadRequestJSONResponse }

func (response DeleteKeyval400JSONResponse) VisitDeleteKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type DeleteKeyval404JSONResponse struct{ NotFoundJSONResponse }

func (response DeleteKeyval404JSONResponse) VisitDeleteKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetKeyvalRequestObject struct {
	Params GetKeyvalParams
}

type GetKeyvalResponseObject interface {
	VisitGetKeyvalResponse(w http.ResponseWriter) error
}

type GetKeyval200JSONResponse KeyVal

func (response GetKeyval200JSONResponse) VisitGetKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetKeyval400JSONResponse struct{ BadRequestJSONResponse }

func (response GetKeyval400JSONResponse) VisitGetKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetKeyval404JSONResponse struct{ NotFoundJSONResponse }

func (response GetKeyval404JSONResponse) VisitGetKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type CreateKeyvalRequestObject struct {
	Body *CreateKeyvalJSONRequestBody
}

type CreateKeyvalResponseObject interface {
	VisitCreateKeyvalResponse(w http.ResponseWriter) error
}

type CreateKeyval201JSONResponse KeyVal

func (response CreateKeyval201JSONResponse) VisitCreateKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateKeyval400JSONResponse struct{ BadRequestJSONResponse }

func (response CreateKeyval400JSONResponse) VisitCreateKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateKeyval409JSONResponse ErrorResponse

func (response CreateKeyval409JSONResponse) VisitCreateKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

type UpdateKeyvalRequestObject struct {
	Body *UpdateKeyvalJSONRequestBody
}

type UpdateKeyvalResponseObject interface {
	VisitUpdateKeyvalResponse(w http.ResponseWriter) error
}

type UpdateKeyval200JSONResponse KeyVal

func (response UpdateKeyval200JSONResponse) VisitUpdateKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type UpdateKeyval400JSONResponse struct{ BadRequestJSONResponse }

func (response UpdateKeyval400JSONResponse) VisitUpdateKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type UpdateKeyval404JSONResponse struct{ NotFoundJSONResponse }

func (response UpdateKeyval404JSONResponse) VisitUpdateKeyvalResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create or update KeyVal asynchroniously
	// (PUT /async)
	PutKeyvalAsync(ctx context.Context, request PutKeyvalAsyncRequestObject) (PutKeyvalAsyncResponseObject, error)
	// Get KeyVal by key from cache (inconsistency possible)
	// (GET /fast)
	GetKeyvalFast(ctx context.Context, request GetKeyvalFastRequestObject) (GetKeyvalFastResponseObject, error)
	// Delete existing KeyVal
	// (DELETE /keyval)
	DeleteKeyval(ctx context.Context, request DeleteKeyvalRequestObject) (DeleteKeyvalResponseObject, error)
	// Get KeyVal by key
	// (GET /keyval)
	GetKeyval(ctx context.Context, request GetKeyvalRequestObject) (GetKeyvalResponseObject, error)
	// Create new KeyVal
	// (POST /keyval)
	CreateKeyval(ctx context.Context, request CreateKeyvalRequestObject) (CreateKeyvalResponseObject, error)
	// Update existing KeyVal
	// (PUT /keyval)
	UpdateKeyval(ctx context.Context, request UpdateKeyvalRequestObject) (UpdateKeyvalResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// PutKeyvalAsync operation middleware
func (sh *strictHandler) PutKeyvalAsync(ctx echo.Context) error {
	var request PutKeyvalAsyncRequestObject

	var body PutKeyvalAsyncJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PutKeyvalAsync(ctx.Request().Context(), request.(PutKeyvalAsyncRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PutKeyvalAsync")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PutKeyvalAsyncResponseObject); ok {
		return validResponse.VisitPutKeyvalAsyncResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetKeyvalFast operation middleware
func (sh *strictHandler) GetKeyvalFast(ctx echo.Context, params GetKeyvalFastParams) error {
	var request GetKeyvalFastRequestObject

	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetKeyvalFast(ctx.Request().Context(), request.(GetKeyvalFastRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetKeyvalFast")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetKeyvalFastResponseObject); ok {
		return validResponse.VisitGetKeyvalFastResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// DeleteKeyval operation middleware
func (sh *strictHandler) DeleteKeyval(ctx echo.Context, params DeleteKeyvalParams) error {
	var request DeleteKeyvalRequestObject

	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteKeyval(ctx.Request().Context(), request.(DeleteKeyvalRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteKeyval")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(DeleteKeyvalResponseObject); ok {
		return validResponse.VisitDeleteKeyvalResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetKeyval operation middleware
func (sh *strictHandler) GetKeyval(ctx echo.Context, params GetKeyvalParams) error {
	var request GetKeyvalRequestObject

	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetKeyval(ctx.Request().Context(), request.(GetKeyvalRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetKeyval")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetKeyvalResponseObject); ok {
		return validResponse.VisitGetKeyvalResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// CreateKeyval operation middleware
func (sh *strictHandler) CreateKeyval(ctx echo.Context) error {
	var request CreateKeyvalRequestObject

	var body CreateKeyvalJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CreateKeyval(ctx.Request().Context(), request.(CreateKeyvalRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateKeyval")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CreateKeyvalResponseObject); ok {
		return validResponse.VisitCreateKeyvalResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// UpdateKeyval operation middleware
func (sh *strictHandler) UpdateKeyval(ctx echo.Context) error {
	var request UpdateKeyvalRequestObject

	var body UpdateKeyvalJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.UpdateKeyval(ctx.Request().Context(), request.(UpdateKeyvalRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "UpdateKeyval")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(UpdateKeyvalResponseObject); ok {
		return validResponse.VisitUpdateKeyvalResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}