# api/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

// Defines values for ErrorResponseCode.
const (
	BADREQUEST   ErrorResponseCode = "BAD_REQUEST"
	KEYVALEXISTS ErrorResponseCode = "KEYVAL_EXISTS"
	NOTFOUND     ErrorResponseCode = "NOT_FOUND"
)

// Valid indicates whether the value is a known member of the ErrorResponseCode enum.
func (e ErrorResponseCode) Valid() bool {
	switch e {
	case BADREQUEST:
		return true
	case KEYVALEXISTS:
		return true
	case NOTFOUND:
		return true
	default:
		return false
	}
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    ErrorResponseCode `json:"code"`
	Message string            `json:"message"`
}

// ErrorResponseCode defines model for ErrorResponse.Code.
type ErrorResponseCode string

// KeyVal defines model for KeyVal.
type KeyVal struct {
	Key   string `json:"key"`
	Value string `json:"value"`
}

// KeyValQuery defines model for KeyValQuery.
type KeyValQuery = string

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// GetKeyvalFastParams defines parameters for GetKeyvalFast.
type GetKeyvalFastParams struct {
	// Key Value key
	Key KeyValQuery `form:"key" json:"key"`
}

// DeleteKeyvalParams defines parameters for DeleteKeyval.
type DeleteKeyvalParams struct {
	// Key Value key
	Key KeyValQuery `form:"key" json:"key"`
}

// GetKeyvalParams defines parameters for GetKeyval.
type GetKeyvalParams struct {
	// Key Value key
	Key KeyValQuery `form:"key" json:"key"`
}

// PutKeyvalAsyncJSONRequestBody defines body for PutKeyvalAsync for application/json ContentType.
type PutKeyvalAsyncJSONRequestBody = KeyVal

// CreateKeyvalJSONRequestBody defines body for CreateKeyval for application/json ContentType.
type CreateKeyvalJSONRequestBody = KeyVal

// UpdateKeyvalJSONRequestBody defines body for UpdateKeyval for application/json ContentType.
type UpdateKeyvalJSONRequestBody = KeyVal